+++ packages/swagen/src/generators/koa.js
@@ -1,6 +1,6 @@
 import path from "path";
 
-import { generate, mkdir, parse } from "../lib";
+import { generate, mkdir, parse, getApiRoles } from "../lib";
 
 /**
  * Generate code for koa server api
@@ -14,26 +14,15 @@ export default function genKoa({
   templatePath,
 }) {
   mkdir(dist);
-  const configPath = path.dirname(dist) + "/config";
-  mkdir(configPath);
   return parse(yamlFile)
     .then(function(swagger) {
       const { api } = swagger;
 
-      // 获取所有api operationId，默认在 dist(./src/api)同目录下(./src/config)生成roles.js
-      const rolePath = configPath + "/roles.js";
-      let roles = [];
-      for (let name in api) {
-        for (let operation of api[name]["operations"]) {
-          roles.push({
-            operationId: operation["operationId"],
-            roles: operation["roles"] || [],
-          });
-        }
-      }
-      // 将 roles 写入 roles.js 文件
+      // 生成 roles.js 文件
       const tplRole = path.join(templatePath, "koa", "role.hbs");
-      generate(tplRole, rolePath, { value: roles });
+      generate(tplRole, path.join(dist, "roles.js"), {
+        value: getApiRoles(api),
+      });
 
       // 生成 api 文件
       const tplAPI = path.join(templatePath, "koa", "api.hbs");
