+++ packages/swagen/templates/koa/api.hbs
@@ -2,7 +2,7 @@
 
 import * as schemas from "./{{name}}.schema.js";
 import { validate, checkRole } from "../middlewares";
-import roles from "../config/roles.js"
+import roles from "./roles.js"
 
 /**
  * @typedef {Object} State
