+++ packages/tpl-service/src/api/pet.js
@@ -2,7 +2,7 @@
 
 import * as schemas from "./pet.schema.js";
 import { validate, checkRole } from "../middlewares";
-import roles from "../config/roles.js";
+import roles from "./roles.js";
 
 /**
  * @typedef {Object} State
