+++ packages/tpl-service/src/api/roles.js
@@ -1,0 +1,9 @@
+const roles = {
+  listPets: ["BREEDER", "OWNER"],
+  createPet: ["BREEDER"],
+  showPetById: ["OWNER"],
+  updatePet: [],
+  deletePet: [],
+};
+
+export default roles;
